pantheon2/backend/requirements.txt
# Pantheon 2.0 Backend Dependencies
# Python 3.11+

# API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Data validation
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Database
supabase>=2.5.0

# LLM providers
anthropic>=0.28.0
openai>=1.35.0
google-generativeai>=0.7.0

# Audio processing (Phase 3)
librosa>=0.10.0
numpy>=1.26.0
sounddevice>=0.4.7

# Local SLM inference (Phase 3)
llama-cpp-python>=0.2.77

# Logging
structlog>=24.1.0

# HTTP / async
httpx>=0.27.0
aiofiles>=23.2.1

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0

# Utilities
python-dotenv>=1.0.0
